Refactor common partitioning/distribution logic to BaseAggregateExec
